Removed the code for specifying the uid/gid
The pilot user is not registered so you should specify a uid or gid.  Let the system decide.

The reason for doing this in the first place was that everything was being run as root and we wanted the pilot to run as an unprivileged account.  This reason is no longer valid since we drop privileges as soon as possible in the bootstrap process.