Added a bad hack to search for the pilot service config file in the same directory as the service binary.
This is because CERN insists on having everything in CVMFS and won't allow cloud images to be contextualized.  This hack assumes that the cloud in question is OpenStack since that is the current whim at CERN.  This is not a portable solution.  The correct solution is to contextualize the image with the appropriate config file, since presumably you know what kind of provider you are uploading the image to.