Added the ability to run pre and post scripts
Depending on the VO and the services or software installed, there may be dependencies that must be satisfied before a pilot can successfully launch.  Adding the ability for a VO to install pre and post scripts to the bootstrap service allows a VO to check for these dependencies and clean up after the pilot exits.

For instance, there pre-packaged pre script that checks for ephemeral storage, mounts it, and moves the home directories to it so that the pilot will have additional space to work in.

The post script section could be a VO specific way to send back logs in case of pilot error.